% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_QAIC_cpp}
\alias{calculate_QAIC_cpp}
\title{Calculate the Quasi-Akaike Information Criterion (QAIC) for a fitted GEE model}
\usage{
calculate_QAIC_cpp(gee_model)
}
\arguments{
\item{gee_model}{A List containing a fitted GEE model object (geeglm) from the `geepack` package, 
with the following elements: 'family', 'fitted.values', 'y', 'residuals', 'coefficients' and 'geese'.}
}
\value{
A numeric value representing the QAIC of the fitted model.
}
\description{
This C++ function calculates the QAIC for a fitted Generalized Estimating Equations (GEE) model.
It supports Gaussian, Binomial, and Poisson family models. The QAIC is used for model comparison in GEE models,
similar to AIC but adjusts for quasi-likelihood estimation and includes a correction for small sample sizes.
}
\details{
The QAIC is calculated using the formula:
\deqn{QAIC = -2 \cdot \text{Quasi-likelihood} + 2 \cdot k + \frac{2 \cdot k \cdot (k + 1)}{n - k - 1}}
where \emph{k} is the number of model parameters, and log-likelihood depends on the model family.
}
\examples{
library(geepack)

# Create example data
set.seed(123)
mydata <- data.frame(
  y = rbinom(100, 1, 0.5),
  x1 = rnorm(100),
  x2 = rnorm(100),
  id = rep(1:10, each = 10)
)

# Fit GEE model
fit_binomial <- geeglm(
  y ~ x1 + x2, family = binomial, data = mydata, 
  id = id, corstr = "exchangeable")

# Calculate QAIC
qaic_value <- calculate_QAIC_cpp(fit_binomial)
print(qaic_value)

}
