% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_QIC_cpp}
\alias{calculate_QIC_cpp}
\title{Calculate the Quasi-likelihood under the Independence model Criterion (QIC) for a fitted GEE model}
\usage{
calculate_QIC_cpp(gee_model)
}
\arguments{
\item{gee_model}{A List containing a fitted GEE model object (geeglm) from the `geepack` package, 
with the following elements: 'family', 'fitted.values', 'y', 'residuals', 'coefficients' and 'geese'.}
}
\value{
A numeric value representing the QIC of the fitted model.
}
\description{
This C++ function calculates the QIC for a fitted Generalized Estimating Equations (GEE) model.
It supports Gaussian, Binomial, and Poisson family models. The QIC is used for model comparison in GEE models,
similar to AIC and BIC.
}
\details{
The QIC is calculated using the formula:
\deqn{QIC = -2 \cdot \text{log-likelihood} + 2 \cdot k}
where \emph{k} is the number of model parameters, and log-likelihood depends on the model family.
}
\examples{
library(geepack)

# Create example data
set.seed(123)
mydata <- data.frame(
  y = rbinom(100, 1, 0.5),
  x1 = rnorm(100),
  x2 = rnorm(100),
  id = rep(1:10, each = 10)
)

# Fit GEE model
fit_binomial <- geeglm(
  y ~ x1 + x2, family = binomial, data = mydata, 
  id = id, corstr = "exchangeable")

# Calculate QIC
qic_value <- calculate_QIC_cpp(fit_binomial)
print(qic_value)

}
