\name{SA24204178-package}
\alias{SA24204178-package}
\docType{package}
\title{This package provides functions for calculating QAIC and QIC for GEE models}
\description{This package offers both R and Rcpp implementations for calculating 
    the Quasi-likelihood under the Independence Model Criterion (QAIC) and the 
    Quasi-likelihood under the Independence model Criterion (QIC) for 
    Generalized Estimating Equations (GEE) models. These functions are designed 
    to assess and compare the fit of GEE models in various regression settings, 
    including Gaussian, binomial, and Poisson families.}
\author{
Ran Deng
Maintainer: Ran Deng <Dr2016@mail.ustc.edu.cn>
}
\keyword{ package }